renamed the createdDate and lastUpdatedDate and lastLoggedInDate for User, Post, Comment
