+++ backend/blog/src/main/java/com/jk/blog/dto/comment/CommentRequestBody.java
@@ -1,4 +1,4 @@
-package com.jk.blog.dto;
+package com.jk.blog.dto.comment;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
