+++ backend/blog/src/main/java/com/jk/blog/dto/profile/ProfileResponseBody.java
@@ -1,7 +1,7 @@
-package com.jk.blog.dto;
+package com.jk.blog.dto.profile;
 
 
-import jakarta.validation.constraints.Pattern;
+import com.jk.blog.entity.User;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,7 +11,9 @@ import java.util.List;
 @Getter
 @Setter
 @NoArgsConstructor
-public class ProfileRequestBody {
+public class ProfileResponseBody {
+
+    private Long profileId;
 
     private String address;
