+++ backend/blog/src/main/java/com/jk/blog/dto/user/UserRequestBody.java
@@ -1,4 +1,4 @@
-package com.jk.blog.dto;
+package com.jk.blog.dto.user;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
@@ -13,8 +13,6 @@ import lombok.Setter;
 @Setter
 public class UserRequestBody {
 
-    private Long id;
-
     @NotEmpty
     @Size(min = 4, message = "Name Should Be Greater Than 4")
     private String name;
@@ -25,11 +23,6 @@ public class UserRequestBody {
 
     @Email(message = "Please Enter Valid Email")
     private String email;
-
-    @NotEmpty(message = "Password is Required")
-    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[\\W_])[A-Za-z\\d\\W_]{8,}$",
-            message = "Password Must Be at least 8 Characters long and include at least one UPPERCASE letter, one lowercase letter, one Number, and one Special Character")
-    private String password;
 
     @NotEmpty(message = "Mobile number must not be empty")
     @Pattern(regexp = "^\\+?[1-9][0-9]{7,14}$", message = "Invalid Mobile Number Format")
